refactor(runner): split monitors out of main
