+++ runner/lib/main.js
@@ -1,7 +1,6 @@
-/* global process setInterval clearInterval console:off */
-/* eslint-disable no-continue */
+/* global process console:off */
 
-import { resolve as resolvePath, join as joinPath, basename } from 'path';
+import { resolve as resolvePath, join as joinPath } from 'path';
 import { performance } from 'perf_hooks';
 import zlib from 'zlib';
 import { promisify } from 'util';
@@ -14,13 +13,7 @@ import yargsParser from 'yargs-parser';
 import chalk from 'chalk';
 import { makePromiseKit } from './sdk/promise-kit.js';
 
-import {
-  sleep,
-  PromiseAllOrErrors,
-  warnOnRejection,
-  aggregateTryFinally,
-  sequential,
-} from './helpers/async.js';
+import { sleep, aggregateTryFinally, sequential } from './helpers/async.js';
 import { childProcessDone } from './helpers/child-process.js';
 import { fsStreamReady, makeFsHelper } from './helpers/fs.js';
 import { makeProcfsHelper } from './helpers/procsfs.js';
@@ -28,6 +21,9 @@ import { makeOutputter } from './helpers/outputter.js';
 
 import { makeTasks as makeLocalChainTasks } from './tasks/local-chain.js';
 import { makeTasks as makeTestnetTasks } from './tasks/testnet.js';
+
+import { makeChainMonitor } from './monitor/chain-monitor.js';
+import { monitorSlog } from './monitor/slog-monitor.js';
 
 /** @typedef {import('./helpers/async.js').Task} Task */
 
@@ -41,102 +37,6 @@ const defaultLoadgenConfig = {
 const defaultMonitorIntervalMinutes = 5;
 const defaultStageDurationMinutes = 6 * 60;
 const defaultNumberStages = 4 + 2;
-
-const vatIdentifierRE = /^(v\d+):(.*)$/;
-
-/**
- * @typedef {{
- *   time: number,
- *   type: 'create-vat',
- *   vatID: string,
- *   name?: string,
- *   dynamic: boolean,
- *   managerType: "local" | "xs-worker" | string,
- * }} SlogCreateVatEvent
- */
-
-/**
- * @typedef {({
- *   time: number,
- *   type: |
- *     'vat-startup-finish' |
- *     'replay-transcript-start' | // Renamed to 'start-replay'
- *     'start-replay',
- *   vatID: string
- * }) | SlogCreateVatEvent} SlogVatEvent
- */
-
-/** @typedef {SlogVatEvent["type"]} SlogVatEventTypes */
-
-/**
- * @typedef { |
- *   'cosmic-swingset-bootstrap-block-start' |
- *   'cosmic-swingset-bootstrap-block-finish' |
- *   'cosmic-swingset-end-block-start' |
- *   'cosmic-swingset-end-block-finish' |
- *   'cosmic-swingset-begin-block'
- * } SlogCosmicSwingsetBlockEventTypes
- */
-
-/**
- * @typedef {{
- *   time: number,
- *   type: SlogCosmicSwingsetBlockEventTypes,
- *   blockHeight?: number,
- *   blockTime: number
- * }} SlogCosmicSwingsetBlockEvent
- */
-
-/**
- * @typedef { |
- *   SlogVatEventTypes |
- *   SlogCosmicSwingsetBlockEventTypes
- * } SlogSupportedEventTypes
- */
-
-/**
- * @typedef { |
- *   SlogVatEvent |
- *   SlogCosmicSwingsetBlockEvent
- * } SlogSupportedEvent
- */
-
-/** @type {SlogVatEventTypes[]} */
-const vatSlogEventTypes = [
-  'create-vat',
-  'vat-startup-finish',
-  'replay-transcript-start',
-  'start-replay',
-];
-
-/** @type {SlogCosmicSwingsetBlockEventTypes[]} */
-const swingsetRegularBlockSlogEventTypes = [
-  'cosmic-swingset-begin-block',
-  'cosmic-swingset-end-block-start',
-  'cosmic-swingset-end-block-finish',
-];
-
-/** @type {SlogCosmicSwingsetBlockEventTypes[]} */
-const swingsetStartupBlockSlogEventTypes = [
-  'cosmic-swingset-bootstrap-block-start',
-  'cosmic-swingset-bootstrap-block-finish',
-];
-
-/** @param {SlogSupportedEventTypes[]} eventTypes */
-const filterSlogEvent = (eventTypes) =>
-  new RegExp(`^{"time":\\d+(?:\\.\\d+),"type":"(?:${eventTypes.join('|')})"`);
-
-const slogEventRE = filterSlogEvent([
-  ...vatSlogEventTypes,
-  ...swingsetRegularBlockSlogEventTypes,
-  ...swingsetStartupBlockSlogEventTypes,
-]);
-
-/**
- * @callback LogPerfEvent
- * @param {string} eventType
- * @param {Record<string, unknown>} [data]
- */
 
 /**
  * @param {unknown} maybeObj
@@ -323,7 +223,7 @@ const main = async (progName, rawArgs, powers) => {
   );
   await fsStreamReady(outputStream);
 
-  /** @type {LogPerfEvent} */
+  /** @type {import('./stats/types.js').LogPerfEvent} */
   const logPerfEvent = (eventType, data = {}) => {
     const perfNowNs = performance.now() * 1000;
     outputStream.write(
@@ -340,380 +240,6 @@ const main = async (progName, rawArgs, powers) => {
       ),
     );
     outputStream.write('\n');
-  };
-
-  /**
-   * @param {Pick<import("./tasks/types.js").RunChainInfo, 'storageLocation' | 'processInfo'>} chainInfo
-   * @param {Object} param1
-   * @param {Console} param1.console
-   * @param {LogPerfEvent} param1.logPerfEvent
-   * @param {number} param1.cpuTimeOffset
-   * @param {import('./helpers/fs.js').DirDiskUsage} param1.dirDiskUsage
-   */
-  const makeChainMonitor = (
-    { storageLocation, processInfo: kernelProcessInfo },
-    // eslint-disable-next-line no-shadow
-    { console, logPerfEvent, cpuTimeOffset, dirDiskUsage },
-  ) => {
-    /**
-     * @typedef {{
-     *    processInfo: import("./helpers/procsfs.js").ProcessInfo | null | undefined,
-     *    vatName: string | undefined,
-     *    local: boolean,
-     *    started: boolean,
-     *  }} VatInfo
-     */
-    /** @type {Map<string, VatInfo>} */
-    const vatInfos = new Map();
-    let vatUpdated = Promise.resolve();
-
-    const updateVatInfos = async () => {
-      console.log('Updating vat infos');
-      const childrenInfos = new Set(
-        await kernelProcessInfo.getChildren().catch(() => []),
-      );
-      for (const info of childrenInfos) {
-        const vatArgv = await info.getArgv(); // eslint-disable-line no-await-in-loop
-        if (!vatArgv || basename(vatArgv[0]).slice(0, 2) !== 'xs') continue;
-        const vatIdentifierMatches = vatIdentifierRE.exec(vatArgv[1]);
-        if (!vatIdentifierMatches) continue;
-        const vatID = vatIdentifierMatches[1];
-        const vatInfo = vatInfos.get(vatID);
-
-        if (!vatInfo) {
-          /** @type {string | undefined} */
-          let vatName = vatIdentifierMatches[2];
-          if (!vatName || vatName === 'undefined') vatName = undefined;
-          // TODO: warn found vat process without create event
-          console.warn(
-            `found vat ${vatID}${
-              vatName ? ` ${vatName}` : ''
-            } process before create event`,
-            'pid=',
-            info.pid,
-          );
-          // vatInfo = { vatName, processInfo: info };
-          // vatInfos.set(vatID, vatInfo);
-          continue;
-        }
-
-        if (vatInfo.processInfo !== info) {
-          // TODO: warn if replacing with new processInfo ?
-        }
-
-        vatInfo.processInfo = info;
-
-        // if (!vatInfo.started) {
-        //   monitorConsole.warn(
-        //     `found vat ${vatID}${
-        //       vatInfo.vatName ? ` ${vatInfo.vatName}` : ''
-        //     } process before vat start event`,
-        //     'pid=',
-        //     info.pid,
-        //   );
-        // }
-      }
-      for (const [vatID, vatInfo] of vatInfos) {
-        if (vatInfo.processInfo && !childrenInfos.has(vatInfo.processInfo)) {
-          vatInfo.processInfo = null;
-        }
-
-        if (vatInfo.started && !vatInfo.local && !vatInfo.processInfo) {
-          // Either the vat started but the process doesn't exist yet (undefined)
-          // or the vat process exited but the vat didn't stop yet (null)
-          console.warn(
-            `Vat ${vatID} started but process ${
-              vatInfo.processInfo === null
-                ? 'exited early'
-                : "doesn't exist yet"
-            }`,
-          );
-        }
-      }
-    };
-
-    const ensureVatInfoUpdated = async () => {
-      const vatUpdatedBefore = vatUpdated;
-      await vatUpdated;
-      if (vatUpdated === vatUpdatedBefore) {
-        vatUpdated = updateVatInfos();
-        warnOnRejection(
-          vatUpdated,
-          console,
-          'Failed to update vat process infos',
-        );
-        await vatUpdated;
-      }
-    };
-
-    const logProcessUsage = async () => {
-      await vatUpdated;
-      const results = await PromiseAllOrErrors(
-        [
-          {
-            eventData: {
-              processType: 'kernel',
-            },
-            processInfo: kernelProcessInfo,
-          },
-          ...[...vatInfos]
-            .filter(([, { local }]) => !local)
-            .map(([vatID, { processInfo, vatName }]) => ({
-              eventData: {
-                processType: 'vat',
-                vatID,
-                name: vatName,
-              },
-              processInfo,
-            })),
-        ].map(async ({ eventData, processInfo }) => {
-          if (!processInfo) {
-            console.warn('missing process', eventData);
-            return false;
-          }
-          const { times, memory } = await processInfo.getUsageSnapshot();
-          logPerfEvent('chain-process-usage', {
-            ...eventData,
-            real:
-              Math.round(
-                performance.now() * 1000 -
-                  (processInfo.startTimestamp - cpuTimeOffset) * 1e6,
-              ) / 1e6,
-            ...times,
-            ...memory,
-          });
-          return true;
-        }),
-      );
-      if (results.some((result) => result === false)) {
-        throw new Error('Missing vat processes.');
-      }
-      return results;
-    };
-
-    const logStorageUsage = async () => {
-      logPerfEvent('chain-storage-usage', {
-        chain: await dirDiskUsage(storageLocation),
-      });
-    };
-
-    /** @type {NodeJS.Timer | null} */
-    let monitorIntervalId = null;
-
-    const stop = () => {
-      if (monitorIntervalId) {
-        clearInterval(monitorIntervalId);
-        monitorIntervalId = null;
-      }
-    };
-
-    /** @param {number} interval */
-    const start = (interval) => {
-      stop();
-      monitorIntervalId = setInterval(
-        () =>
-          warnOnRejection(
-            PromiseAllOrErrors([logStorageUsage(), logProcessUsage()]),
-            console,
-            'Failure during usage monitoring',
-          ),
-        interval,
-      );
-    };
-
-    /**
-     *
-     * @param {string} vatID
-     * @param {string | undefined} vatName
-     * @param {string} vatType
-     */
-    const createVat = (vatID, vatName, vatType) => {
-      if (!vatInfos.has(vatID)) {
-        vatInfos.set(vatID, {
-          vatName,
-          processInfo: undefined,
-          local: vatType === 'local',
-          started: false,
-        });
-      } else {
-        // TODO: warn already created vat before
-      }
-    };
-
-    /**
-     *
-     * @param {string} vatID
-     */
-    const updateVat = (vatID) => {
-      const vatInfo = vatInfos.get(vatID);
-      if (!vatInfo) {
-        // TODO: warn unknown vat
-      } else if (!vatInfo.processInfo) {
-        ensureVatInfoUpdated();
-      }
-    };
-
-    return harden({
-      start,
-      stop,
-      logProcessUsage,
-      logStorageUsage,
-      createVat,
-      updateVat,
-    });
-  };
-
-  /**
-   * @param {Pick<import("./tasks/types.js").RunChainInfo, 'slogLines'>} chainInfo
-   * @param {Object} param1
-   * @param {() => void} param1.resolveFirstEmptyBlock
-   * @param {ReturnType<makeChainMonitor>} [param1.chainMonitor]
-   * @param {LogPerfEvent} param1.logPerfEvent
-   * @param {import("stream").Writable} [param1.slogOutput]
-   * @param {Console} param1.console
-   */
-  const monitorSlog = async (
-    { slogLines },
-    // eslint-disable-next-line no-shadow
-    { resolveFirstEmptyBlock, chainMonitor, logPerfEvent, slogOutput, console },
-  ) => {
-    /** @type {number | null}  */
-    let slogStart = null;
-
-    let slogBlocksSeen = 0;
-    let slogEmptyBlocksSeen = 0;
-    let slogLinesInBlock = 0;
-
-    for await (const line of slogLines) {
-      if (slogOutput) {
-        slogOutput.write(line);
-      }
-
-      if (slogStart == null && chainMonitor) {
-        // TODO: figure out a better way
-        // There is a risk we could be late to the party here, with the chain
-        // having started some time before us but in reality we usually find
-        // the process before it starts the kernel
-        slogStart = performance.now() / 1000;
-        warnOnRejection(
-          chainMonitor.logStorageUsage(),
-          console,
-          'Failed to get first storage usage',
-        );
-      }
-
-      slogLinesInBlock += 1;
-
-      // Avoid JSON parsing or converting lines we don't care about
-      // Parse as ascii, in case the payload has multi-byte chars,
-      // the time and type tested prefix is guaranteed to be single-byte.
-      if (!slogEventRE.test(line.toString('ascii', 0, 100))) continue;
-
-      const localEventTime = performance.timeOrigin + performance.now();
-
-      /** @type {SlogSupportedEvent} */
-      let event;
-      try {
-        event = JSON.parse(line.toString('utf8'));
-      } catch (error) {
-        console.warn('Failed to parse slog line', line, error);
-        continue;
-      }
-
-      const delay = Math.round(localEventTime - event.time * 1000);
-
-      if (delay > 100) {
-        console.log('slog event', event.type, 'delay', delay, 'ms');
-      }
-
-      switch (event.type) {
-        case 'create-vat': {
-          if (chainMonitor) {
-            const { vatID, name: vatName, managerType } = event;
-            chainMonitor.createVat(vatID, vatName, managerType);
-          }
-          break;
-        }
-        case 'vat-startup-finish':
-        case 'replay-transcript-start':
-        case 'start-replay': {
-          if (chainMonitor) {
-            const { vatID } = event;
-            chainMonitor.updateVat(vatID);
-          }
-          break;
-        }
-        case 'cosmic-swingset-bootstrap-block-start': {
-          logPerfEvent('chain-first-init-start');
-          break;
-        }
-        case 'cosmic-swingset-bootstrap-block-finish': {
-          logPerfEvent('chain-first-init-finish');
-          break;
-        }
-        case 'cosmic-swingset-begin-block': {
-          const { blockHeight = 0 } = event;
-          if (!slogBlocksSeen) {
-            logPerfEvent('stage-first-block', { block: blockHeight });
-            if (chainMonitor) {
-              await chainMonitor.logProcessUsage().catch((usageErr) => {
-                // Abuse first empty block as it will be awaited before monitorChain
-                // And won't abruptly end our monitor
-                // @ts-ignore resolving with a rejected promise is still "void" ;)
-                resolveFirstEmptyBlock(Promise.reject(usageErr));
-              });
-            }
-          }
-          console.log('begin-block', blockHeight);
-          slogBlocksSeen += 1;
-          break;
-        }
-        case 'cosmic-swingset-end-block-start': {
-          if (event.blockHeight === 0) {
-            // Before https://github.com/Agoric/agoric-sdk/pull/3491
-            // bootstrap didn't have it's own slog entry
-            logPerfEvent('chain-first-init-start');
-          } else {
-            slogLinesInBlock = 0;
-          }
-          break;
-        }
-        case 'cosmic-swingset-end-block-finish': {
-          if (event.blockHeight === 0) {
-            // TODO: measure duration from start to finish
-            logPerfEvent('chain-first-init-finish');
-          } else {
-            const { blockHeight = 0 } = event;
-            // Finish line itself doesn't count
-            slogLinesInBlock -= 1;
-            if (slogLinesInBlock === 0) {
-              if (!slogEmptyBlocksSeen) {
-                logPerfEvent('stage-first-empty-block', {
-                  block: blockHeight,
-                });
-                resolveFirstEmptyBlock();
-              }
-              slogEmptyBlocksSeen += 1;
-            }
-
-            console.log(
-              'end-block',
-              blockHeight,
-              'linesInBlock=',
-              slogLinesInBlock,
-            );
-          }
-          break;
-        }
-        default: {
-          console.warn(
-            'Parsed an unexpected slog event:',
-            // @ts-expect-error
-            event.type,
-          );
-        }
-      }
-    }
   };
 
   /**
