+++ runner/lib/stats/types.d.ts
@@ -1,0 +1,5 @@
+/* eslint-disable no-unused-vars,no-redeclare */
+
+export interface LogPerfEvent {
+  (eventType: string, data?: Record<string, unknown>): void;
+}
